refactor: Reception 을 Controller 로 이동
